medical_records_service/requirements.txt
Django>=5.0,<6
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
